fix: fix migrations types

+++ database/migrations/1732382272260_create_aeds_table.ts
@@ -12,7 +12,7 @@ export default class extends BaseSchema {
       table.text("address_line2").nullable();
       table.text("photo_url").nullable();
 
-      table.bigInteger("building_id").unsigned().nullable();
+      table.integer("building_id").unsigned().nullable();
       table
         .foreign("building_id")
         .references("buildings.id")
